fix(proto,device): spec-correct extended property services wire format
`A_PropertyExt*` services were parsed and encoded with the wrong wire
layout. Per KNX spec 03_03_07 §3.4.3.2 / Figure 49 the
`object_instance` and `property_id` fields are each **12 bits** and
share octets 4-6 of the APDU:

    [4]:   object_instance[11..4]
    [5]:   object_instance[3..0] | property_id[11..8]
    [6]:   property_id[7..0]

The old parser treated octets 4-5 as a full 16-bit `object_instance`
(BE) and octet 6 as an 8-bit `property_id`, leaving the low 4 bits
of `object_instance` overlapping the high 4 bits of `property_id`.
Wire bytes `00 10 0C` decoded as `(instance=0x0010, pid=0x0C)`
instead of the spec-correct `(instance=1, pid=0x00C)`. The
`SystemBObjects` dispatch compensated by subtracting `0x0010 - 1`
before resolving — an artifact the conformance templates were
written around.

This also broke negative tests in suites 4.1.4 / 4.2.4 / 4.5.4 /
4.5.7 / 4.6.4 / 4.7.4 / 3.8.7.3: they inject wire bytes like
`00 18 00 01 ...` intending `(instance=1, pid=0x800)` to exercise
"non-existing PID returns `E_ADDRESS_VOID`". Under the old parser
those bytes decoded as `(instance=0x0018, pid=0)` and either
silently dropped the frame or failed to round-trip the response.

### Changes

- `PropertyExtValueHeader` / `FunctionPropertyExtHeader`: parse and
  write the 12/12-bit packed layout via new
  `unpack_instance_and_pid` / `pack_instance_and_pid` helpers.
  `prop_id` widens from `u8` to `u16`; `object_instance` stays `u16`
  but now carries the correctly unpacked 12-bit value.
- `handle_ext_description_read`: same unpacking in the inline parse
  and response writer.
- `InterfaceObject::resolve_object_index`: widened to `u16` so the
  extended-services 12-bit instance range maps through without
  truncation. Default `resolve_ext_object_index` now just delegates;
  the SystemB override is deleted. The cEMI Local Management
  caller (`device_mgmt.rs`) stays 8-bit on the wire and widens at
  the boundary.
- `InterfaceObjectAugment` dispatch, `PropertyLookup::ByPid`,
  `FullPropertyReadRequest.pid`, `FullPropertyWriteRequest.pid`,
  `FunctionPropertyRequest.prop_id`, `PropertyDescriptor.pid`,
  `PropertyDescriptionResponse.prop_id` — all `u8` → `u16` to carry
  the wire value end-to-end. PID constants in
  `objects/interface/mod.rs` widened accordingly.
- Regular (non-extended) `PropertyValueHeader` /
  `PropertyDescriptionRead` / `FunctionPropertyHeader` field types
  also widen to `u16` to share the `InterfaceObjects` trait surface;
  their wire writers cast back to `u8` at the byte boundary since
  the regular services really do carry an 8-bit PID per spec
  03_03_07 §3.4.3.1.
- `A_SystemNetworkParameterResponse` keeps its 8-bit PID — spec
  03_05_02 §2.20 defines the field that way — with a cast at the
  call site.

### Verification

- `cargo test -p zweidraehte-proto --lib apdu` — 63 pass, new tests
  cover the full 12-bit PID round-trip and the conformance
  `00 10 0C → (instance=1, pid=0x0C)` reference decode.
- `cargo run --bin conformance-runner` (secure DUT): **508/508**
  pass, up from 502/508 on the branch baseline. The six previously-
  failing tests (`4.1.4`, `4.2.4`, `4.5.4`, `4.5.7`, `4.6.4`,
  `4.7.4`, `3.8.7.3`) all pass now.
- `cargo run --bin conformance-runner -- --non-secure` (plain DUT):
  **193/193** pass — no regression.

SESSION.md: dropped the stale "Object Instance Convention" note
(it described the workaround, not the bug) and added a follow-up
note to rewrite `handle_ext_description_read` using a typed
`PropertyExtDescriptionHeader` analogous to the other APDU
headers.

+++ crates/zweidraehte-device/src/bcus/system_b/extensions/ip/augment.rs
@@ -206,7 +206,7 @@ const KNXNETIP_CAP_TUNNELING_BIT: u16 = 1 << 1;
 // ============================================================================
 
 // Helper to build a simple non-array property descriptor.
-const fn simple_desc(pid: u8, pdt_id: u8, access: PropertyAccess) -> PropertyDescriptor {
+const fn simple_desc(pid: u16, pdt_id: u8, access: PropertyAccess) -> PropertyDescriptor {
     let (read_level, write_level) = match access {
         PropertyAccess::ReadOnly | PropertyAccess::WriteOnly => (3, 0),
         PropertyAccess::ReadWrite => (3, 3),
@@ -260,7 +260,7 @@ impl<P: IpPlatform, const N: usize, const CAPS: u16> IpAugment<'_, P, N, CAPS> {
     }
 
     /// Look up a property descriptor by PID.
-    fn ip_descriptor_by_pid(&self, prop_id: u8) -> Option<PropertyDescriptor> {
+    fn ip_descriptor_by_pid(&self, prop_id: u16) -> Option<PropertyDescriptor> {
         if self.tunneling_enabled() {
             let max_addrs = N as u16;
             match prop_id {
