+++ crates/zweidraehte-device/src/layers/application/services/system_network_parameter.rs
@@ -80,7 +80,7 @@ fn handle_read<D: StackDefinition>(ind: &KnxMessageBuffer<Buffer<'static>>, ctx:
     // Only the serial-number-by-programming-mode procedure is supported.
     // Per spec §2.20, unsupported parameter_type/test_info combinations
     // MUST NOT trigger a response.
-    if object_type != device_ot || pid_val != pid::SERIAL_NUMBER || operand != OPERAND_BY_PROG_MODE {
+    if object_type != device_ot || u16::from(pid_val) != pid::SERIAL_NUMBER || operand != OPERAND_BY_PROG_MODE {
         trace!(
             "AL SystemNetworkParameterRead unsupported: object_type=0x{:03X}, pid={}, operand=0x{:02X}",
             object_type, pid_val, operand
@@ -119,7 +119,15 @@ fn handle_read<D: StackDefinition>(ind: &KnxMessageBuffer<Buffer<'static>>, ctx:
     .build();
 
     let serial: &[u8; 6] = ctx.state.serial_number();
-    SystemNetworkParameterResponse::write(msg.buf_mut(), device_ot, pid::SERIAL_NUMBER, OPERAND_BY_PROG_MODE, serial);
+    // pid::SERIAL_NUMBER fits in 8 bits; SystemNetworkParameter wire
+    // encoding uses an 8-bit PID field (spec 03_05_02 §2.20).
+    SystemNetworkParameterResponse::write(
+        msg.buf_mut(),
+        device_ot,
+        pid::SERIAL_NUMBER as u8,
+        OPERAND_BY_PROG_MODE,
+        serial,
+    );
 
     ctx.lctx.push_outbox(msg.into_inner());
 }
